SDF file scanning fixes.  No more Molecule class.
Squashed commit of the following:

commit ee68546b3c735550907580b20e3e1eae6987da6a
Author: Justin Gullingsrud <gullingj@deshawresearch.com>
Date:   Mon Nov 30 08:52:35 2015 -0500

    Sanitize ut to use tempfile.


commit 5a219dd8862f3729ebe0acc14e900222108b522c
Author: Justin Gullingsrud <gullingj@deshawresearch.com>
Date:   Sun Nov 29 11:11:44 2015 -0500

    Fix incorrect rounding in sdf coord output.


commit c4801f6983902dffd352189fa4c13d8096466efd
Author: Justin Gullingsrud <gullingj@deshawresearch.com>
Date:   Sun Nov 29 08:57:53 2015 -0500

    Fix sdf entry name regression.


commit d09c9dd15160351f89afe67f480c19705ceb8004
Author: Justin Gullingsrud <gullingj@deshawresearch.com>
Date:   Sun Nov 29 08:32:46 2015 -0500

    Bring back FormatSDF.


commit 6a2f7898a880ee59ac7ba007f5da2d8cd0f759ac
Author: Justin Gullingsrud <gullingj@deshawresearch.com>
Date:   Fri Nov 27 15:41:01 2015 -0500

    Factor out analyze_residue.


commit 875889d044dd8c241fb043684045e42cf8c405f5
Author: Justin Gullingsrud <gullingj@deshawresearch.com>
Date:   Fri Nov 27 15:36:47 2015 -0500

    Static initializers for maps.


commit 9908a5d1fa5916e7a215d6ecd5d2eaa8fae6e3b4
Author: Justin Gullingsrud <gullingj@deshawresearch.com>
Date:   Fri Nov 27 15:26:11 2015 -0500

    Factor System::analyze() out into Analyze(System)


commit ba37ef3b3198ed2e8bcd31818abcc13b9168dfb3
Author: Justin Gullingsrud <gullingj@deshawresearch.com>
Date:   Fri Nov 27 15:04:32 2015 -0500

    Split SystemImporter into its own file.


commit 01344bd0078633ac710a4357e65e71ec9a13fdfa
Author: Justin Gullingsrud <gullingj@deshawresearch.com>
Date:   Fri Nov 27 14:37:50 2015 -0500

    Make FromSmilesString create System.

    No more Molecule class.


commit dd32cff68bef2b687759c90ad208b7d40e3a5dc4
Author: Justin Gullingsrud <gullingj@deshawresearch.com>
Date:   Fri Nov 27 14:16:28 2015 -0500

    Unit tests pass.


commit 9d63c0ca04324a326bd4aeeffc5594c4b5bf05cd
Author: Justin Gullingsrud <gullingj@deshawresearch.com>
Date:   Fri Nov 27 11:22:43 2015 -0500

    Purge serialiation.  More sdf I/O changes.

    Got rid of tools/info/inchi as well.


commit bde8b46da3aa6ec1b69e119cf447223f601962e3
Author: Justin Gullingsrud <gullingj@deshawresearch.com>
Date:   Thu Nov 26 20:09:35 2015 -0500

    Remove testArchive, since archive went away.


commit 86106449e21710802b61b89146232d062ee63f95
Author: Justin Gullingsrud <justinrocks@gmail.com>
Date:   Thu Nov 26 20:04:15 2015 -0500

    Expand System to take over Molecule and Annotated.

+++ src/pdb/readpdb.h
@@ -142,7 +142,7 @@ void desres_msys_get_pdb_cryst1(const char *record,
 static void desres_msys_get_pdb_coordinates(const char *record, 
                                 double *x, double *y, double *z,
                                 double *occup, double *beta,
-                                int *formal_charge) {
+                                int8_t *formal_charge) {
   char numstr[60]; /* store all fields in one array to save memset calls */
   memset(numstr, 0, sizeof(numstr));
 
@@ -297,7 +297,7 @@ static void desres_msys_get_pdb_fields(const char *record, int reclength, int *s
                            char *altloc, char *elementsymbol,
                            double *x, double *y, double *z, 
                            double *occup, double *beta,
-                           int* formal_charge) {
+                           int8_t* formal_charge) {
   char serialbuf[6];
 
   /* get atom serial number */
